fix: fixed landing page text overflow, (#16)
feat: add logout feature on landing page
