+++ components/header/DashboardHeader.jsx
@@ -10,27 +10,16 @@ import axios from 'axios';
 import { useRouter } from 'next/navigation';
 import LanguageMenu from './LanguageMenu';
 import useLanguageStore from '@/lib/zustand/useLanguageStore';
+import AvatarButton from './AvatarButton';
 
 const DashboardHeader = ({ onOpenHandler }) => {
-  const { user, setIsLogin, setUser } = useUserStore((state) => state);
-  const [isAvatarMenuOpen, setIsAvatarMenuOpen] = useState(false);
+  const { setIsLogin, setUser } = useUserStore((state) => state);
   const [isLangMenuOpen, setIsLangMenuOpen] = useState(false);
   const { setLang } = useLanguageStore();
 
   const router = useRouter();
 
-  const toggleAvatarMenu = () => setIsAvatarMenuOpen(!isAvatarMenuOpen);
   const toggleLangMenu = () => setIsLangMenuOpen(!isLangMenuOpen);
-
-  useEffect(() => {
-    const handleClickOutside = (e) => {
-      if (!e.target.closest('.avatar-menu')) setIsAvatarMenuOpen(false);
-    };
-
-    document.addEventListener('click', handleClickOutside);
-
-    return () => document.removeEventListener('click', handleClickOutside);
-  }, []);
 
   useEffect(() => {
     const handleClickOutside = (e) => {
@@ -40,36 +29,11 @@ const DashboardHeader = ({ onOpenHandler }) => {
     return () => document.removeEventListener('click', handleClickOutside);
   }, []);
 
-  const logoutHandler = async () => {
-    try {
-      const data = await axios.post(
-        `${process.env.NEXT_PUBLIC_API_URL}/logout`,
-        {},
-        {
-          withCredentials: true,
-        }
-      );
-
-      if (data.status === 200) {
-        setIsLogin(false);
-        setUser({});
-
-        router.push('/');
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
   return (
     <header className='sticky w-full z-50 top-0 left-0 bg-neutral-100'>
       <div className='flex justify-between px-4 py-3 border-b border-border lg:flex-row-reverse'>
         <div className='lg:hidden'>
-          <Button
-            onClick={onOpenHandler}
-            className='rounded-full bg-white w-10 border'
-            variant='reverse'
-          >
+          <Button onClick={onOpenHandler} className='rounded-full bg-white w-10 border' variant='reverse'>
             <RxHamburgerMenu />
           </Button>
         </div>
@@ -77,19 +41,7 @@ const DashboardHeader = ({ onOpenHandler }) => {
           <div className='lang-menu'>
             <LanguageMenu isOpen={isLangMenuOpen} toggle={toggleLangMenu} setLang={setLang} variant='reverse' />
           </div>
-          <div className='relative avatar-menu'>
-            <Avatar className='cursor-pointer' onClick={toggleAvatarMenu}>
-              <AvatarImage src={user.picture} />
-              <AvatarFallback>AV</AvatarFallback>
-            </Avatar>
-            <ul className={`absolute top-[calc(100%+0.75rem)] right-0 transform transition-all duration-300 ${isAvatarMenuOpen ? 'opacity-100 scale-100 translate-y-0' : 'opacity-0 scale-95 pointer-events-none translate-y-[-10px]'} z-50`}>
-              <Card className='bg-white py-2 flex flex-col shadow-md'>
-                <li className='px-6 py-2 hover:bg-black hover:text-white cursor-pointer' onClick={logoutHandler}>
-                  <span className='text-lg'>Logout</span>
-                </li>
-              </Card>
-            </ul>
-          </div>
+          <AvatarButton />
         </div>
       </div>
     </header>
